Add files via upload
In this project we want to predict the position on the left-right scale
of Polish respondents. In the previous projects we have shown that there
was a significant difference between Polish populist radical right
parties (PRRP) voters and others on their attitude towards immigration
and religiosity. Even if the results were significant, we could not
exclude the possibility that third parties, such as socio-demographic
variables, had an influence on the results. Furthermore, these results
do not yet help us to predict PRRP voters' personality based on their attitudes. To
fill these gaps, we will use a linear regression analysis in this
section.